split helpers module into smaller modules with unit tests
