+++ src/masonite/utils/collections.py
@@ -520,3 +520,16 @@ class Collection:
             items = items.all()
 
         return items
+
+
+def flatten(iterable):
+    """Flatten all sub-iterables of an iterable structure (recursively)."""
+    flat_list = []
+    for item in iterable:
+        if isinstance(item, list):
+            for subitem in flatten(item):
+                flat_list.append(subitem)
+        else:
+            flat_list.append(item)
+
+    return flat_list
