+++ tests/core/response/test_response_download.py
@@ -1,9 +1,7 @@
 from tests import TestCase
-from src.masonite.utils.helpers import generate_wsgi
 from src.masonite.foundation import Application
 import os
 from src.masonite.response import Response
-from src.masonite.routes import Router, Route
 
 
 class TestResponseRedirect(TestCase):
