+++ tests/core/response/test_response_redirections.py
@@ -1,5 +1,4 @@
 from tests import TestCase
-from src.masonite.utils.helpers import generate_wsgi
 from src.masonite.foundation import Application
 import os
 from src.masonite.response import Response
